Cambios 0.1.5

+++ src/Listeners/MeetingEventSubscriber.php
@@ -2,20 +2,19 @@
 
 namespace RoiUp\Zoom\Listeners;
 
- use RoiUp\Zoom\Events\Meeting\MeetingCreated;
- use RoiUp\Zoom\Events\Meeting\MeetingDeleted;
- use RoiUp\Zoom\Events\Meeting\MeetingUpdated;
- use RoiUp\Zoom\Events\Meeting\MeetingStarted;
- use RoiUp\Zoom\Events\Meeting\MeetingEnded;
- use RoiUp\Zoom\Events\Notifications\SendDeleteOccurrence;
- use RoiUp\Zoom\Events\User\UserVerified;
- use RoiUp\Zoom\Models\Eloquent\Host;
- use RoiUp\Zoom\Models\Eloquent\Meeting;
- use RoiUp\Zoom\Models\Eloquent\Meeting as EloquentModel;
- use RoiUp\Zoom\Models\Eloquent\Occurrence;
- use RoiUp\Zoom\Models\Eloquent\Registrant as RegistrantModel;
+use RoiUp\Zoom\Events\Meeting\MeetingCreated;
+use RoiUp\Zoom\Events\Meeting\MeetingDeleted;
+use RoiUp\Zoom\Events\Meeting\MeetingEnded;
+use RoiUp\Zoom\Events\Meeting\MeetingStarted;
+use RoiUp\Zoom\Events\Meeting\MeetingUpdated;
+use RoiUp\Zoom\Events\Notifications\SendDeleteOccurrence;
+use RoiUp\Zoom\Events\User\UserVerified;
+use RoiUp\Zoom\Models\Eloquent\Host;
+use RoiUp\Zoom\Models\Eloquent\Meeting;
+use RoiUp\Zoom\Models\Eloquent\Meeting as EloquentModel;
+use RoiUp\Zoom\Models\Eloquent\Occurrence;
 
- class MeetingEventSubscriber extends AbstractEventSubscriber
+class MeetingEventSubscriber extends AbstractEventSubscriber
 {
 
     private $actions = ['Created', 'Updated', 'Deleted', 'Started', 'Ended'];
@@ -25,22 +24,36 @@ namespace RoiUp\Zoom\Listeners;
      */
     public function onMeetingCreated(MeetingCreated $event) {
 
-        $this->logEvent($event);
 
         if($event->getObject()['type'] === 4){
             $host = Host::whereHostId($event->getObject()['host_id'])->whereInvitationStatus('pending')->first();
             if(!empty($host)){
+                $this->logEvent($event);
+
                 $host->invitation_status = 'accepted';
                 $host->save();
                 event(new UserVerified($host));
+                $this->logFinishEvent();
+            }else{
+                $this->logNotFoundEvent();
             }
         }else{
             $model = Meeting::whereZoomId($event->getObject()['id'])->first();
-            $model->status = 'created';
-            $model->save();
+            if(!empty($model)){
+
+                $this->logEvent($event);
+
+                $model->status = 'created';
+                $model->save();
+
+                $this->logFinishEvent();
+            }else{
+                $this->logNotFoundEvent();
+            }
+
         }
 
-        $this->logFinishEvent();
+
     }
 
     /**
@@ -48,35 +61,40 @@ namespace RoiUp\Zoom\Listeners;
      */
     public function onMeetingDeleted(MeetingDeleted $event) {
 
-        $this->logEvent($event);
-
         $meeting = Meeting::whereZoomId($event->getObject()['id'])->first();
+        if(!empty($meeting)){
+            $this->logEvent($event);
+            $occurrences = isset($event->getObject()['occurrences']) ? $event->getObject()['occurrences'] : $meeting->occurrences;
 
-        $occurrences = isset($event->getObject()['occurrences']) ? $event->getObject()['occurrences'] : $meeting->occurrences;
+            foreach($occurrences as $occurrence){
 
-        foreach($occurrences as $occurrence){
+                $item = Occurrence::whereOccurrenceId($occurrence['occurrence_id'])->whereMeetingId($meeting->zoom_id)->first();
 
-            $item = Occurrence::whereOccurrenceId($occurrence['occurrence_id'])->whereMeetingId($meeting->zoom_id)->first();
+                if($item == null){
+                    continue;
+                }
 
-            if($item == null){
-                continue;
+                $item->registrants->each(function($registrant){
+                    if($registrant->status !== 'denied'){
+                        event(new SendDeleteOccurrence($registrant));
+                    }
+                    $registrant->delete();
+                });
+
+                $item->delete();
             }
 
-            $item->registrants->each(function($registrant){
-                if($registrant->status !== 'denied'){
-                    event(new SendDeleteOccurrence($registrant));
-                }
-                $registrant->delete();
-            });
+            if(Occurrence::whereMeetingId($meeting->zoom_id)->count() == 0){
+                $meeting->delete();
+            }
 
-            $item->delete();
+            $this->logFinishEvent();
+        }else{
+            $this->logNotFoundEvent();
         }
 
-        if(Occurrence::whereMeetingId($meeting->zoom_id)->count() == 0){
-            $meeting->delete();
-        }
 
-        $this->logFinishEvent();
+
     }
 
     /**
@@ -84,39 +102,42 @@ namespace RoiUp\Zoom\Listeners;
      */
     public function onMeetingUpdated(MeetingUpdated $event) {
 
-        $this->logEvent($event);
-
         $meeting = EloquentModel::whereZoomId($event->getObject()['id'])->first();
+        if(!empty($meeting)) {
 
-        $changes = $event->getObject();
+            $this->logEvent($event);
 
-        unset($changes['id']);
-        foreach($changes as $field => $value){
+            $changes = $event->getObject();
 
-            switch ($field){
-                case 'recurrence':
-                    $meeting->$field = json_encode($value);
-                    break;
-                case 'settings':
-                    $meeting->mergeSettings($value);
-                    break;
-                case 'occurrences':
-                    $meeting->updateOccurrences($value);
-                    break;
-                default:
-                    if(in_array($field, $meeting->getFillable())){
-                        $meeting->$field = $value;
-                    }
-                    break;
+            unset($changes['id']);
+            foreach ($changes as $field => $value) {
+
+                switch ($field) {
+                    case 'recurrence':
+                        $meeting->$field = json_encode($value);
+                        break;
+                    case 'settings':
+                        $meeting->mergeSettings($value);
+                        break;
+                    case 'occurrences':
+                        $meeting->updateOccurrences($value);
+                        break;
+                    default:
+                        if (in_array($field, $meeting->getFillable())) {
+                            $meeting->$field = $value;
+                        }
+                        break;
+                }
             }
+
+            $meeting->duration = $meeting->occurrences[0]['duration'];
+            $meeting->start_time = $meeting->occurrences[0]['start_time'];
+
+            $meeting->save();
+            $this->logFinishEvent();
+        }else{
+            $this->logNotFoundEvent();
         }
-
-        $meeting->duration = $meeting->occurrences[0]['duration'];
-        $meeting->start_time = $meeting->occurrences[0]['start_time'];
-
-        $meeting->save();
-
-        $this->logFinishEvent();
     }
 
     /**
@@ -124,9 +145,8 @@ namespace RoiUp\Zoom\Listeners;
      */
     public function onMeetingStarted(MeetingStarted $event) {
 
-        $this->logEvent($event);
-        $this->changeMeetingStatus($event->getObject()['id'], 'started');
-        $this->logFinishEvent();
+        $this->changeMeetingStatus($event->getObject()['id'], 'started', $event);
+
     }
 
     /**
@@ -134,13 +154,24 @@ namespace RoiUp\Zoom\Listeners;
      */
     public function onMeetingEnded(MeetingEnded $event) {
 
-        $this->logEvent($event);
-        $this->changeMeetingStatus($event->getObject()['id'], 'ended');
-        $this->logFinishEvent();
+        $this->changeMeetingStatus($event->getObject()['id'], 'ended', $event);
+
     }
 
-    private function changeMeetingStatus($meetingId, $status){
-        Meeting::whereZoomId($meetingId)->update(['status' => $status]);
+    private function changeMeetingStatus($meetingId, $status, $event){
+        $model = Meeting::whereZoomId($meetingId)->first();
+        if(!empty($model)){
+
+            $this->logEvent($event);
+
+            $model->status = $status;
+            $model->save();
+
+            $this->logFinishEvent();
+
+        }else{
+            $this->logNotFoundEvent();
+        }
     }
 
     /**
